#82 Add email authentication feature - Add nodemailer dependency - Add nodemailer transport file - Add email authentication function
