+++ server/config/emailAuth.template.ts
@@ -1,0 +1,12 @@
+import * as nodemailer from 'nodemailer';
+
+export const smtpTransport = nodemailer.createTransport({
+  service: '',
+  auth: {
+    user: '',
+    pass: '',
+  },
+  tls: {
+    // rejectUnauthorized: boolean,
+  },
+});
