+++ server/package.json
@@ -23,6 +23,7 @@
     "ip": "^1.1.8",
     "jsonwebtoken": "^8.5.1",
     "mysql2": "^2.3.3",
+    "nodemailer": "^6.7.8",
     "os": "^0.1.2",
     "reflect-metadata": "^0.1.13",
     "sequelize": "^6.21.3",
@@ -38,6 +39,7 @@
     "@types/ip": "^1.1.0",
     "@types/jsonwebtoken": "^8.5.9",
     "@types/node": "^18.6.4",
+    "@types/nodemailer": "^6.4.5",
     "@types/swagger-jsdoc": "^6.0.1",
     "@types/swagger-ui-express": "^4.1.3",
     "@types/validator": "^13.7.5",
